+++ spec/spec_helper.rb
@@ -16,6 +16,9 @@ ENV["RAILS_ENV"] = "test"
 db_file = File.expand_path("dummy/db/test.sqlite3", __dir__)
 File.unlink(db_file) if File.exist?(db_file)
 
+# Needed for loading Rails 6.x and 7.0
+require "logger"
+
 Bundler.require(:default, :test)
 
 require "dotenv/load"