+++ spec/super_settings/storage/redis_storage_spec.rb
@@ -85,7 +85,7 @@ if EXTENSIONS[:redis]
       it "should return the last time a setting was updated" do
         setting_1 = SuperSettings::Storage::RedisStorage.new(key: "setting_1", raw_value: "1", updated_at: Time.now - 100)
         setting_1.save!
-        setting_2 = SuperSettings::Storage::RedisStorage.new(key: "setting_2", raw_value: "2", updated_at: Time.at(Time.now - 50.to_i))
+        setting_2 = SuperSettings::Storage::RedisStorage.new(key: "setting_2", raw_value: "2", updated_at: Time.at(Time.now - 50))
         setting_2.save!
         setting_2 = SuperSettings::Storage::RedisStorage.find_by_key(setting_2.key)
         expect(SuperSettings::Storage::RedisStorage.last_updated_at).to eq setting_2.updated_at